# Conflicts:
#	LiDARC-backend/src/main/java/com/example/lidarcbackend/service/comparisons/ComparisonService.java
#	LiDARC-frontend/src/app/components/comparison-details/comparison-details.ts

+++ LiDARC-backend/src/main/java/com/example/lidarcbackend/service/comparisons/ComparisonService.java
@@ -249,7 +249,7 @@ public class ComparisonService implements IComparisonService {
     }
 
     private ComparisonPlan processFolderGroup(List<Long> fileIds, GridParameters grid, Comparison savedComparison,
-                                              String groupName, Integer pointFilterLowerBound, Integer pointFilterUpperBound,
+                                              String groupName, Double pointFilterLowerBound, Double pointFilterUpperBound,
                                               Boolean outlierDetectionEnabled, Double outlierDeviationFactor, Boolean needPointFilter)
         throws NotFoundException {
         ComparisonPlan plan = new ComparisonPlan();
